requirements.txt
playwright>=1.40.0
cryptography>=41.0.0
Pillow>=10.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
pyyaml>=6.0
requests>=2.31.0
click>=8.1.0
rich>=13.0.0
